+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `exec` command
 - [server] upgrade golang version of server to 1.9
 - [helm] Add HealthChecks
+- [server] graceful shutdown
 
 ## [0.14.0] - 2018-02-06
 ### Changed